requirements.txt
aio-pika==9.5.8
aiormq==6.9.2
anyio==4.11.0
asgiref==3.10.0
async-timeout==5.0.1
certifi==2025.11.12
channels==4.3.1
charset-normalizer==3.4.4
click==8.3.0
Django==4.2.7
django-cors-headers==4.3.1
django-prometheus==2.3.1
django-redis==5.4.0
djangorestframework==3.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
pamqp==3.3.0
pika==1.3.2
prometheus-client==0.19.0
propcache==0.4.1
psycopg==3.2.12
psycopg-binary==3.2.12
python-dateutil==2.8.2
python-decouple==3.8
pytz==2025.2
redis==7.0.1
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
wheel==0.45.1
yarl==1.22.0
